PFE/requirements.txt
# requirements.txt

# Core
streamlit>=1.28.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# AI/ML
openai>=0.28.0
transformers>=4.35.0
torch>=2.0.0
spacy>=3.6.0

# CV Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
plotly>=5.17.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Authentication
bcrypt>=4.0.0
jwt>=1.3.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0

# For Arabic support
arabic-reshaper>=3.0.0
python-bidi>=0.4.0

# Monitoring
streamlit-option-menu>=0.3.0
